requirements.txt
# Core Dependencies

# AI & Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
ultralytics==8.0.207
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
scipy==1.11.4

# Web Framework
flask==3.0.0
flask-cors==4.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0

# Communication & Messaging
paho-mqtt==1.6.1
firebase-admin==6.2.0
requests==2.31.0

# Audio & Signal Processing
librosa==0.10.0
soundfile==0.12.1
scipy==1.11.4
pyaudio==0.2.13

# GPS & Location
gpxpy==1.6.2
geopy==2.3.0
polyline==2.0.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
logging-config==0.5.2
colorama==0.4.6
tqdm==4.66.1

# Hardware/IoT
pyserial==3.5
adafruit-circuitpython-ads1x15==2.0.4

# Data Science & Analysis
pandas==2.1.1
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
plotly==5.18.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
mock==5.1.0

# Code Quality
pylint==3.0.3
flake8==6.1.0
black==23.12.0
isort==5.13.2

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Development
ipython==8.18.1
jupyter==1.0.0
notebook==7.0.6
ipdb==0.13.13
